+++ src/main/java/ru/yandex/practicum/filmorate/dal/mappers/MpaRowMapper.java
@@ -12,7 +12,7 @@ public class MpaRowMapper implements RowMapper<Mpa> {
     @Override
     public Mpa mapRow(ResultSet rs, int rowNum) throws SQLException {
         Mpa mpa = new Mpa();
-        mpa.setId(rs.getLong("mpa_id"));
+        mpa.setId(rs.getInt("mpa_id"));
         mpa.setName(rs.getString("name"));
         return mpa;
     }
